Moved normalization code into LDNN because it is specific to LDNN and not submodels in general. This also makes its use automatic for train and test.
Added an optimized version of gradient descent for a batch size of 1 without dropout which will be used by GLIA eventually (looks to be about 8x faster).

Seperated out the shuffle Cython function into its own module and made it so that it no longer needs a special process to be imported.