Added some leniency to the multi-level K-means algorithm RMS checking.
This now will allow some small 'rounding errors' to occur when calculating the next round of means causing an increase in the RMS^2 value instead of decreasing like it always should. This seems to always happen with the nucleoli supplemental data and IQR normalization it has been seen in other situations as well (e.g. when running 50 iterations for timing sometimes 1 would die this way). The errors must occur below the main level and be at most 0.5% of the RMS^2 value.